splitting pg - in progress.
