+++ index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+require('dotenv').config();
+
+require('./server/server.js').start(process.env.PORT);
+
