+++ routes/routes.js
@@ -1,38 +1,5 @@
 'use strict';
 
-require('dotenv').config();
-
-// Application Dependencies
-const express = require('express');
-const pg = require('pg');
-const superagent = require('superagent');
-const methodOverride = require('method-override');
-
-// Application Setup
-const app = express();
-const PORT = process.env.PORT;
-
-// Database Setup
-const client = new pg.Client(process.env.DATABASE_URL);
-client.connect();
-client.on('error', err => console.error(err));
-
-// Application Middleware
-app.use(express.urlencoded({ extended: true }));
-app.use(express.static('public'));
-
-app.use(methodOverride((request, response) => {
-  if (request.body && typeof request.body === 'object' && '_method' in request.body) {
-    // look in urlencoded POST bodies and delete it
-    let method = request.body._method;
-    delete request.body._method;
-    return method;
-  }
-}))
-
-// Set the view engine for server-side templating
-app.set('view engine', 'ejs');
-
 // API Routes
 app.get('/', getBooks);
 app.post('/searches', createSearch);
@@ -44,7 +11,17 @@ app.delete('/books/:id', deleteBook);
 
 app.get('*', (request, response) => response.status(404).send('This route does not exist'));
 
-app.listen(PORT, () => console.log(`Listening on port: ${PORT}`));
+module.exports = app;
+
+
+
+const pg = require('pg');
+const superagent = require('superagent');
+
+// Database Setup
+const client = new pg.Client(process.env.DATABASE_URL);
+client.connect();
+client.on('error', err => console.error(err));
 
 // HELPER FUNCTIONS
 function Book(info) {
@@ -169,3 +146,5 @@ function deleteBook(request, response) {
 function handleError(error, response) {
   response.render('pages/error', { error: error });
 }
+
+module.exports = {getBooks, createSearch, newSearch, getBook, createShelf, createBook, updateBook, deleteBook}
